# project/public/swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/notifications": {
      "get": {
        "description": "Get all notifications for authenticated user",
        "operationId": "NotificationController_findAll",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Notifications retrieved successfully"
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get all notifications",
        "tags": [
          "Notifications"
        ]
      }
    },
    "/notifications/{id}/read": {
      "put": {
        "description": "Mark a specific notification as read",
        "operationId": "NotificationController_markAsRead",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Notification ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Notification marked as read successfully"
          },
          "404": {
            "description": "Notification not found"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Mark notification as read",
        "tags": [
          "Notifications"
        ]
      }
    },
    "/notifications/read-all": {
      "put": {
        "description": "Mark all notifications as read for authenticated user",
        "operationId": "NotificationController_markAllAsRead",
        "parameters": [],
        "responses": {
          "200": {
            "description": "All notifications marked as read successfully"
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Mark all notifications as read",
        "tags": [
          "Notifications"
        ]
      }
    },
    "/notifications/{id}": {
      "delete": {
        "description": "Delete a specific notification",
        "operationId": "NotificationController_deleteNotification",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Notification ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Notification deleted successfully"
          },
          "404": {
            "description": "Notification not found"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Delete notification",
        "tags": [
          "Notifications"
        ]
      }
    },
...
    "/rescue-teams": {
      "post": {
        "description": "Create a new rescue team organization",
        "operationId": "RescueController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateRescueTeamDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Rescue team created successfully"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Create rescue team",
        "tags": [
          "Rescue Teams"
        ]
      },
      "get": {
        "description": "Get list of all rescue teams with optional search",
        "operationId": "RescueController_findAll",
        "parameters": [
          {
            "name": "search",
            "required": false,
            "in": "query",
            "description": "Search query for rescue team name or location",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Rescue teams retrieved successfully"
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get all rescue teams",
        "tags": [
          "Rescue Teams"
        ]
      }
    },
    "/rescue-teams/{id}": {
      "get": {
        "description": "Get detailed information about a specific rescue team",
        "operationId": "RescueController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Rescue team ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Rescue team retrieved successfully"
          },
          "404": {
            "description": "Rescue team not found"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get rescue team by ID",
        "tags": [
          "Rescue Teams"
        ]
      },
      "put": {
        "description": "Update rescue team information",
        "operationId": "RescueController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Rescue team ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateRescueTeamDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Rescue team updated successfully"
          },
          "404": {
            "description": "Rescue team not found"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Update rescue team",
        "tags": [
          "Rescue Teams"
        ]
      }
    },
    "/rescue-teams/{id}/status": {
      "put": {
        "description": "Update the availability status of a rescue team",
        "operationId": "RescueController_updateStatus",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Rescue team ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateRescueTeamStatusDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Status updated successfully"
          },
          "404": {
            "description": "Rescue team not found"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Update rescue team status",
        "tags": [
          "Rescue Teams"
        ]
      }
    },
    "/rescue-teams/available": {
      "get": {
        "description": "Find available rescue teams within a specified radius",
        "operationId": "RescueController_findAvailableTeams",
        "parameters": [
          {
            "name": "latitude",
            "required": true,
            "in": "query",
            "description": "Latitude coordinate",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "longitude",
            "required": true,
            "in": "query",
            "description": "Longitude coordinate",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "radius",
            "required": false,
            "in": "query",
            "description": "Search radius in kilometers (default: 10)",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Available rescue teams retrieved successfully"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Find available rescue teams",
        "tags": [
          "Rescue Teams"
        ]
      }
    },
...
  },
  "info": {
    "title": "Emergency Project API",
    "description": "\n      ## Emergency Response Management System API\n      \n      This API provides comprehensive endpoints for managing emergency requests, hospitals, rescue teams, and related services.\n      \n      ### Features:\n      - **Authentication**: OAuth 2.0 (Google, Facebook, Apple) and Email/Password\n      - **Emergency Requests**: Create and manage SOS emergency requests\n      - **Hospital Management**: Hospital capacity, emergency acceptance\n      - **Rescue Teams**: Team management and assignment\n      - **Dashboard**: Real-time statistics and monitoring\n      - **Notifications**: Real-time push notifications\n      - **Reports**: Generate and download reports\n      - **User Settings**: Manage user preferences and settings\n      \n      ### Authentication:\n      Most endpoints require JWT Bearer token authentication. Use the /auth endpoints to obtain tokens.\n    ",
    "version": "1.0",
    "contact": {
      "name": "Emergency Project Team",
      "url": "https://emergency-project.com",
      "email": "support@emergency-project.com"
    }
  },
...
  "servers": [
    {
      "url": "http://localhost:3001",
      "description": "Local Development Server"
    },
    {
      "url": "https://api.emergency-project.com",
      "description": "Production Server"
    }
  ],
  "components": {
    "securitySchemes": {
      "JWT-auth": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "name": "JWT",
        "description": "Enter JWT token",
        "in": "header"
      }
    },
    "schemas": {
...
      "CreateRescueTeamDto": {
        "type": "object",
        "properties": {}
      },
      "UpdateRescueTeamDto": {
        "type": "object",
        "properties": {}
      },
      "UpdateRescueTeamStatusDto": {
        "type": "object",
        "properties": {}
      },
...
    }
  }
}